requirements.txt
django==4.1.4
django-mptt==0.14.0
djangorestframework==3.14.0
pillow==9.3.0
psycopg2-binary==2.9.5
python-decouple==3.6
